Create table-sort.mjs
Exports the code as an ECMAScript module and stops calling the sort function on page load. In a web application, as opposed to a browser extension, this should be left to developers when to call sorting. There are scenarios with no initial tables (tables are loaded dynamically and the sort function is called when required). As for file extension, it can be ".mjs", which is the standard or alternatively, something like ".esm.js"